requirements.txt
praw
flask
transformers
torch
finnhub-python
numpy<2
requests
python-dotenv